@@ -1,10 +1,10 @@
-const WebpackAssetsManifest = require('webpack-assets-manifest');
+import { WebpackAssetsManifest } from 'webpack-assets-manifest';
 
 const manifest = new WebpackAssetsManifest({
   output: 'merged-manifest.json',
   merge: true,
   customize(entry, original, manifest, asset) {
-    if ( manifest.isMerging ) {
+    if (manifest.isMerging) {
       // Do something
     }
   },
